[U] 64bit対応(主にlong->int変更)、実機動作確認OK

+++ app/src/main/jni/Popup/PopupSrch.h
@@ -15,7 +15,7 @@ struct TPopupForWordsParam {
 	const tchar *word;
 	const tchar *prevword;
 	int click_pos;	// relative position from word.
-	FlexArray<ulong> wordlengths;
+	FlexArray<uint> wordlengths;
 	const tchar *curp;
 	int curi;
